requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorclass==2.2.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.10.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
dnspython==2.1.0
docopt==0.6.2
easygui==0.98.2
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
msoffcrypto-tool==4.12.0
mypy-extensions==0.4.3
oauthlib==3.1.1
olefile==0.46
oletools==0.60
parso==0.8.2
pathspec==0.8.1
pcodedmp==1.2.6
pdfminer==20191125
pipreqs==0.4.10
prompt-toolkit==3.0.18
psycopg2==2.9.1
ptpython==3.0.17
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pyftpdlib==1.5.6
Pygments==2.8.1
PyJWT==2.2.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dotenv==0.19.1
python3-openid==3.2.0
pythonwhois==2.4.3
pytz==2021.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.11
sqlparse==0.4.2
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.4
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.3.0
win-unicode-console==0.5
yarg==0.1.9
